+++ SAM/positioning.h
@@ -32,7 +32,7 @@ enum {
 void UpdateNonActorRefr();
 void SaveObjectTranslation();
 
-void AdjustObjectPosition(int type, double scalar, int step);
+void AdjustObjectPosition(int type, int scalar, int step);
 void ResetObjectPosition();
 void ResetObjectRotation();
 void ResetObjectScale();
